pyproject.toml
[tool.poetry]
name = "reminder-bot"
version = "0.1.0"
description = "telegram bot for reminders"
authors = ["rdcm <rdcm@protonmail.ch>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = { version = "^21.1.1", extras = ["job-queue"] }
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
poetry = "^1.8.2"
black = "^24.4.0"
pylint = "^3.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

